+++ CHANGELOG.md
@@ -18,7 +18,6 @@ All clusters created with a module version older than `v1.0.0-beta.10` need to b
 ## Deprecations
 
 - The `core_services_config.storage` variable is deprecated in favour of the `storage` variable, this will be removed in the `v1.16.0` release.
-- The `lnrs.io/local-storage` node label is deprecated, this will be removed in the `v1.16.0` release.
 - The `logging.control_plane.log_analytics.external_workspace` variable is deprecated as the Log Analytics Workspace for control plane logs should be created outside the module and passed in via the `logging.control_plane.log_analytics.workspace_id`, this will be removed in `v1.18.0` and to use Log Analytics for control plane logs the workspace ID will need passing in via the `logging.control_plane.log_analytics.workspace_id` input variable.
 - The `logging.control_plane.storage_account.id` variable is deprecated in favor of the `logging.storage_account_config.id` input variable. It will be removed in `v1.18.0`. Storage accounts for both workload and control plane logs can be configured via the `logging.storage_account_config.id` input variable.
 
@@ -37,6 +36,8 @@ All clusters created with a module version older than `v1.0.0-beta.10` need to b
 ## [v1.16.0] - UNRELEASED
 
 ### All Changes
+
+- Removed the `lnrs.io/local-storage` node label which was deprecated. [@hadeeds](https://github.com/hadeeds)
 
 ## [v1.15.0] - 2023-06-22